Move parse and build to their own source files
